Auto correct system using edit distance algorithm
We made edit function and implemented methods like insert, replace and delete to be used by edit function. Data used is a novel by shakespeare and is uploaded with the code itself. This edit function only takes possible words that are one or two edits away from the word we input and gives those recommendations with their probability.